Fix issues after merge
